refactor: Participants 생성자 수정
