+++ src/main/java/nextstep/ladders/controller/LadderController.java
@@ -3,25 +3,13 @@ package nextstep.ladders.controller;
 import nextstep.ladders.domain.Ladder;
 import nextstep.ladders.domain.Participants;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
 public class LadderController {
 
-    public static final String REGEX_COMMA = ",";
-
     public Participants register(final String participantsText) {
-        List<String> names = parseParticipants(participantsText);
-        return new Participants(names);
+        return new Participants(participantsText);
     }
 
     public Ladder start(final int countOfPerson, final String maxLadderHeightText) {
         return new Ladder(countOfPerson, maxLadderHeightText);
-    }
-
-    private List<String> parseParticipants(final String participantsText) {
-        return Arrays.stream(participantsText.split(REGEX_COMMA))
-                .collect(Collectors.toList());
     }
 }
